refactor: 仓库重整 — 清除冗余、中英文双语规范化
核心变更:
- DataAggregator 封装互斥锁 (lock/unlock), 消除 MotorTask/Tasks 重复代码
- FollowLogic 直接返回 MotorCmd struct, 删除字符串协议往返和 parseFollowCmd()
- VisionBridge 用 strtol/strtof+endptr 代替 const_cast 消除 UB
- DataAggregator::getJson() 改用 snprintf 固定 buffer 避免 heap 碎片
- MotorDriver 使用 Config.h MOTOR_DEAD_TIME_MS 代替本地重复常量
- 删除死代码: turnCommand(), computeOffset(), parseFollowCmd(), 静态 lockAggregator
- 所有注释/文档 → 中文 + 英文专业名词 (PWM, FreeRTOS, UART, YOLO 等保持英文)
- 新增 Test/ArduinoMock.h 共享测试 mock, 测试不再内联副本
- 删除 docs/MULTI_AGENT_TASKS.md (非项目文档)
- README 修复不存在的文件引用, 添加 VIS 协议说明

+++ ESP32_Solo/ProtocolUtils.cpp
@@ -1,11 +1,11 @@
 #include "ProtocolUtils.h"
 
-// Zero-allocation overload for const char*
+// ─── const char* 版本 (零 heap 分配) ───
 bool verifyChecksum(const char* line, int prefixLen) {
     const char* star = strchr(line, '*');
-    int starPos = star ? (star - line) : -1;
+    int starPos = star ? (int)(star - line) : -1;
     if (starPos <= prefixLen) {
-        // No '*' or '*' too early: backward-compatible, accept
+        // 无 '*' 或 '*' 位置过早 → 向后兼容，直接通过
         return true;
     }
     uint8_t checksum = 0;
@@ -16,10 +16,11 @@ bool verifyChecksum(const char* line, int prefixLen) {
     return checksum == expected;
 }
 
+// ─── String 版本 ───
 bool verifyChecksum(const String& line, int prefixLen) {
     int starPos = line.indexOf('*');
     if (starPos <= prefixLen) {
-        // No '*' or '*' too early: backward-compatible, accept
+        // 无 '*' 或 '*' 位置过早 → 向后兼容，直接通过
         return true;
     }
     uint8_t checksum = 0;
